refactor: remove obsolete admin monitoring and management tests
