--- citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/admin/AdminDataSummarySystemTest.java
@@ -1,57 +1,0 @@
-package com.hack23.cia.systemintegrationtest.admin;
-
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.openqa.selenium.WebElement;
-
-import com.hack23.cia.systemintegrationtest.categories.IntegrationTest;
-import com.hack23.cia.web.impl.ui.application.views.common.menufactory.api.PageCommandAdminConstants;
-
-// ...imports...
-
-@Category(IntegrationTest.class)
-public class AdminDataSummarySystemTest extends AbstractAdminTest {
-
-	@Test(timeout = DEFAULT_TIMEOUT)
-	public void verifyDataManagementOperationsTest() throws Exception {
-		pageVisit.visitDirectPage(PageCommandAdminConstants.COMMAND_DATASUMMARY);
-
-		// Test refresh views
-		final WebElement refreshButton = pageVisit.findButton("Refresh Views");
-		assertNotNull(refreshButton);
-		pageVisit.performClickAction(refreshButton);
-
-		// Test update search index
-		final WebElement indexButton = pageVisit.findButton("Update Search Index");
-		assertNotNull(indexButton);
-		pageVisit.performClickAction(indexButton);
-	}
-
-	@Test(timeout = DEFAULT_TIMEOUT)
-	public void verifyDataManagementRemoveOperationsTest() throws Exception {
-		pageVisit.visitDirectPage(PageCommandAdminConstants.COMMAND_DATASUMMARY);
-
-		assertNotNull("Expect to find Remove Application History button",
-				pageVisit.findButton("Remove Application History"));
-		assertNotNull("Expect to find Remove Documents button", pageVisit.findButton("Remove Documents"));
-		assertNotNull("Expect to find Remove Politicians button", pageVisit.findButton("Remove Politicians"));
-	}
-
-	@Test(timeout = DEFAULT_TIMEOUT)
-	public void verifyAuthorDataSummaryTest() throws Exception {
-		pageVisit.visitDirectPage(PageCommandAdminConstants.COMMAND_AUTHOR_DATASUMMARY);
-		pageVisit.verifyPageContent("Author Summary");
-		pageVisit.validatePage(PageCommandAdminConstants.COMMAND_AUTHOR_DATASUMMARY);
-	}
-
-	@Test(timeout = DEFAULT_TIMEOUT)
-	public void verifyRefreshViewsTest() throws Exception {
-		pageVisit.visitDirectPage(PageCommandAdminConstants.COMMAND_DATASUMMARY);
-		final WebElement refreshViewsButton = pageVisit.findButton("Refresh Views");
-		assertNotNull("Expect to find a Refresh Views Button", refreshViewsButton);
-		pageVisit.performClickAction(refreshViewsButton);
-	}
-
-	
-	// ...other data management tests...
-}
