--- citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/admin/AdminLanguageTest.java
@@ -1,21 +1,0 @@
-package com.hack23.cia.systemintegrationtest.admin;
-
-import org.junit.Test;
-
-import com.hack23.cia.systemintegrationtest.ui.TestConstants;
-import com.hack23.cia.web.impl.ui.application.views.admin.AdminViewConstants;
-import com.hack23.cia.web.impl.ui.application.views.common.menufactory.api.PageCommandAdminConstants;
-
-public final class AdminLanguageTest extends AbstractAdminTest {
-
-    @Test(timeout = TestConstants.DEFAULT_TIMEOUT)
-    public void shouldDisplayLanguageManagement() throws Exception {
-        pageVisit.visitDirectPage(PageCommandAdminConstants.COMMAND_LANGUAGE);
-        
-        verifyViewContent(
-            AdminViewConstants.ADMIN_LANGUAGE_MANAGEMENT,
-            AdminViewConstants.LANGUAGE_OVERVIEW,
-            AdminViewConstants.LANGUAGE_MANAGEMENT_DESCRIPTION
-        );
-    }
-}
